Fix sizeof test, convert type enum into UInt32

+++ SharpObjects.Model/DataObjectValueTypeExtensions.cs
@@ -8,8 +8,8 @@ namespace SharpObjects.Model
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		internal static Boolean HasFlagFast(this DataObjectValue.DataObjectValueType source, DataObjectValue.DataObjectValueType flag)
 		{
-			var sourceValue = (UInt64)source;
-			var flagValue = (UInt64)flag;
+			var sourceValue = (UInt32)source;
+			var flagValue = (UInt32)flag;
 
 			return (sourceValue & flagValue) == flagValue;
 		}
@@ -17,8 +17,8 @@ namespace SharpObjects.Model
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		internal static Boolean HasAnyOfFlags(this DataObjectValue.DataObjectValueType source, DataObjectValue.DataObjectValueType flags)
 		{
-			var sourceValue = (UInt64)source;
-			var flagsValue = (UInt64)flags;
+			var sourceValue = (UInt32)source;
+			var flagsValue = (UInt32)flags;
 
 			return (sourceValue & flagsValue) != 0;
 		}
